Modified drawing to use the raw LED data directly.
This saves a lot of computation as one does not have to map from the
virtual string to the real one with every pixel write. Also we only use
as much CPU as we really need to. We still might run out of frame time
when using really long LED strings, but not having to do the mapping
calculation for every pixel will not happen on every LED so it will
still be better.